Extract `auto_update_ui` crate
