+++ crates/auto_update_ui/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "auto_update"
+name = "auto_update_ui"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,23 +9,20 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/auto_update.rs"
-doctest = false
+path = "src/auto_update_ui.rs"
 
 [dependencies]
 anyhow.workspace = true
+auto_update.workspace = true
 client.workspace = true
-db.workspace = true
+editor.workspace = true
 gpui.workspace = true
 http_client.workspace = true
-log.workspace = true
-paths.workspace = true
+markdown_preview.workspace = true
+menu.workspace = true
 release_channel.workspace = true
-schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-settings.workspace = true
 smol.workspace = true
-tempfile.workspace = true
-which.workspace = true
+util.workspace = true
 workspace.workspace = true
