+++ crates/auto_update_ui/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
