+++ crates/auto_update_ui/src/update_notification.rs
